split out instruction memory locations methods
