+++ tests/unit/compiler/venom/test_mem_ssa.py
@@ -9,8 +9,13 @@ from vyper.venom.analysis.mem_ssa import (
     MemoryPhi,
     MemoryUse,
 )
-from vyper.venom.basicblock import EMPTY_MEMORY_ACCESS, IRBasicBlock, IRLabel
+from vyper.venom.basicblock import IRBasicBlock, IRLabel
 from vyper.venom.effects import Effects
+from vyper.venom.memory_location import (
+    EMPTY_MEMORY_ACCESS,
+    get_read_memory_location,
+    get_write_memory_location,
+)
 
 
 @pytest.fixture
@@ -555,8 +560,8 @@ def test_analyze_instruction_with_no_memory_ops(create_mem_ssa):
     assignment_inst = bb.instructions[0]  # %1 = 42
 
     # Verify that the instruction doesn't have memory operations
-    assert assignment_inst.get_read_memory_location() is EMPTY_MEMORY_ACCESS
-    assert assignment_inst.get_write_memory_location() is EMPTY_MEMORY_ACCESS
+    assert get_read_memory_location(assignment_inst) is EMPTY_MEMORY_ACCESS
+    assert get_write_memory_location(assignment_inst) is EMPTY_MEMORY_ACCESS
 
     assert mem_ssa.memalias.alias_sets is not None
